Update generateTime.ts - Handle noon use case
Hello @SherAtrium

I have been giving your app to my team members for a while now.  I love it.  I do see a bug in how it handles the noon hours on 12 hr setting.  Instead of 12 it says 00.  I added an if to account for this.  I hope you don't mind.